refactor: actualizar rutas de búsqueda de películas y series; eliminar código no utilizado y optimizar manejo de estados
