+++ src/components/Series/Index.jsx
@@ -2,7 +2,7 @@
 import "./series.scss";
 import LayoutMovieSection from "@/layouts/Layout";
 import Button from "@/components/Buttons/Button";
-import Search from "@/components/SearchInput/Search";
+// import Search from "@/components/SearchInput/Search";
 import Container from "@/components/LoadingContainer/Container";
 import Error from "@/components/Error/Error";
 import { motion } from "framer-motion";
@@ -10,17 +10,15 @@ import { useState, useEffect } from "react";
 import dynamic from "next/dynamic";
 import { container, item } from "./animation";
 
+const MediaCardDynamic = dynamic(
+  () => import("@/components/MediaCard/MediaCard"),
+  {
+    loading: () => <Container />,
+  }
+);
+
 export default function Series() {
-  // estados de data series y busqueda
   const [serieData, setSerieData] = useState([]);
-  const [search, setSearch] = useState("");
-  const [dataSearch, setDataSearch] = useState([]);
-
-  // estados de filtros
-  const [valueGender, setValueGender] = useState("");
-  const [genderFiltered, setGenderFiltered] = useState([]);
-
-  // estados de UX
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const [nextPageSerie, setNext] = useState(() => {
@@ -33,14 +31,14 @@ export default function Series() {
 
   useEffect(() => {
     const getSeries = async () => {
+      setLoading(true);
       try {
         const response = await fetch(`/api/series?page=${nextPageSerie}`);
         const data = await response.json();
         setSerieData(data);
-        setLoading(false);
-        setError(null);
       } catch (error) {
         setError(error);
+      } finally {
         setLoading(false);
       }
     };
@@ -48,48 +46,9 @@ export default function Series() {
   }, [nextPageSerie]);
 
   useEffect(() => {
-    const getSearch = async () => {
-      try {
-        const response = await fetch(`/api/series/searcher?value=${search}`);
-        const data = await response.json();
-        setDataSearch(data);
-        setLoading(false);
-        setError(null);
-      } catch (error) {
-        setError(error);
-        setLoading(false);
-      }
-    };
-    getSearch();
-  }, [search]);
-
-  useEffect(() => {
-    const getGenderFiltered = async () => {
-      try {
-        const response = await fetch(
-          `/api/series/filtergender?value=${valueGender}&page=${nextPageSerie}`
-        );
-        const data = await response.json();
-        setGenderFiltered(data);
-        setLoading(false);
-        setError(null);
-      } catch (error) {
-        setError(error);
-        setLoading(false);
-      }
-    };
-    getGenderFiltered();
-  }, [valueGender, nextPageSerie]);
-
-  useEffect(() => {
     sessionStorage.setItem("currentPageSerie", nextPageSerie.toString());
   }, [nextPageSerie]);
 
-  if (error) {
-    return <Error message={"ocurrio un error de parte de nosotros :("} />;
-  }
-
-  // busqueda de paginas --> + 1
   const handlerNextMovie = () => {
     setNext(nextPageSerie + 1);
   };
@@ -98,67 +57,14 @@ export default function Series() {
     setNext(nextPageSerie - 1);
   };
 
-  // atrapa el valor de search y lo setea en el estado
-  const handlerSearch = (e) => {
-    setSearch(e.target.value);
-  };
-
-  // Función de búsqueda de películas
-
-  let result = [];
-  if (!search && !valueGender) {
-    // Si no hay término de búsqueda ni género seleccionado, mostrar todas las películas
-    result = serieData;
-  } else if (!search && valueGender) {
-    // Si no hay término de búsqueda pero hay un género seleccionado, mostrar películas filtradas por género
-    result = genderFiltered.results;
-  } else if (search && !valueGender) {
-    // Si hay un término de búsqueda pero no hay género seleccionado, mostrar resultados de búsqueda
-    result = dataSearch.results.filter((serie) =>
-      serie.name.toLowerCase().includes(search.toLowerCase())
-    );
-  } else {
-    // Si hay término de búsqueda y también un género seleccionado, aplicar ambos filtros
-    result = dataSearch.results.filter(
-      (serie) =>
-        serie.name.toLowerCase().includes(search.toLowerCase()) &&
-        serie.genre_ids.includes(valueGender)
-    );
+  if (error) {
+    return <Error message={"ocurrio un error de parte de nosotros :("} />;
   }
 
-  const handleButtonClick = (id) => {
-    setValueGender(id);
-  };
-
-  const handlerCloseSearch = () => {
-    setSearch("");
-  };
-
-  // media card dynamic
-  const MediaCardDynamic = dynamic(
-    () => import("@/components/MediaCard/MediaCard"),
-    {
-      loading: () => <Container />,
-      ssr: false,
-    }
-  );
+  console.log("SERIES", serieData);
 
   return (
     <LayoutMovieSection>
-      {/* buscador  */}
-      {/* <div className="searcherSerie">
-        <Search
-          funtion={handlerSearch}
-          filter={handleButtonClick}
-          value={search}
-          close={handlerCloseSearch}
-        />
-      </div> */}
-
-      {/* <div className="series_title">
-        <h3>Series</h3>
-      </div> */}
-
       {/* contedor de peliculas */}
       <motion.div
         className="contenedor"
@@ -166,7 +72,7 @@ export default function Series() {
         initial="hidden"
         animate="visible"
       >
-        {result?.map((serie) => (
+        {serieData?.map((serie) => (
           <motion.div variants={item} key={serie.id}>
             <MediaCardDynamic data={serie} />
           </motion.div>
